Limited Access to Controllers Using ACL in Laravel

+++ app/Http/Controllers/API/UserController.php
@@ -130,6 +130,7 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
         $user->delete();
         return ['message' => 'User Deleted'];
